+++ releasenotes.md
@@ -28,6 +28,7 @@
 - Struct members declared in a single line declaration were not sharing attributes. #1266
 - `opt` project setting now properly documented.
 - Incorrect justify formatting of integers.
+- Assertion with duplicate function pointer signatures #1286
 
 ### Stdlib changes